requirements.txt
streamlit
pycryptodome
aiofiles
plotly
watchdog
pandas
graphviz